chore: removed all migrations in favor of one
