--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1681363155_collections_snapshot.js
@@ -1,182 +1,0 @@
-migrate((db) => {
-  const snapshot = [
-    {
-      "id": "3fhw2mfr9zrgodj",
-      "created": "2022-10-03 21:50:44.238Z",
-      "updated": "2023-04-13 05:19:15.647Z",
-      "name": "hooks",
-      "type": "base",
-      "system": false,
-      "schema": [
-        {
-          "system": false,
-          "id": "j8mewfur",
-          "name": "collection",
-          "type": "text",
-          "required": true,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "4xcxcfuv",
-          "name": "event",
-          "type": "select",
-          "required": true,
-          "unique": false,
-          "options": {
-            "maxSelect": 1,
-            "values": [
-              "insert",
-              "update",
-              "delete"
-            ]
-          }
-        },
-        {
-          "system": false,
-          "id": "u3bmgjpb",
-          "name": "action_type",
-          "type": "select",
-          "required": true,
-          "unique": false,
-          "options": {
-            "maxSelect": 1,
-            "values": [
-              "command",
-              "post"
-            ]
-          }
-        },
-        {
-          "system": false,
-          "id": "kayyu1l3",
-          "name": "action",
-          "type": "text",
-          "required": true,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "zkengev8",
-          "name": "action_params",
-          "type": "text",
-          "required": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "balsaeka",
-          "name": "expands",
-          "type": "text",
-          "required": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "emgxgcok",
-          "name": "disabled",
-          "type": "bool",
-          "required": false,
-          "unique": false,
-          "options": {}
-        }
-      ],
-      "indexes": [
-        "CREATE INDEX `_3fhw2mfr9zrgodj_created_idx` ON `hooks` (`created`)"
-      ],
-      "listRule": null,
-      "viewRule": null,
-      "createRule": null,
-      "updateRule": null,
-      "deleteRule": null,
-      "options": {}
-    },
-    {
-      "id": "_pb_users_auth_",
-      "created": "2023-04-13 05:09:04.237Z",
-      "updated": "2023-04-13 05:19:15.645Z",
-      "name": "users",
-      "type": "auth",
-      "system": false,
-      "schema": [
-        {
-          "system": false,
-          "id": "users_name",
-          "name": "name",
-          "type": "text",
-          "required": false,
-          "unique": false,
-          "options": {
-            "min": null,
-            "max": null,
-            "pattern": ""
-          }
-        },
-        {
-          "system": false,
-          "id": "users_avatar",
-          "name": "avatar",
-          "type": "file",
-          "required": false,
-          "unique": false,
-          "options": {
-            "maxSelect": 1,
-            "maxSize": 5242880,
-            "mimeTypes": [
-              "image/jpeg",
-              "image/png",
-              "image/svg+xml",
-              "image/gif",
-              "image/webp"
-            ],
-            "thumbs": null
-          }
-        }
-      ],
-      "indexes": [
-        "CREATE INDEX `__pb_users_auth__created_idx` ON `users` (`created`)"
-      ],
-      "listRule": "id = @request.auth.id",
-      "viewRule": "id = @request.auth.id",
-      "createRule": "",
-      "updateRule": "id = @request.auth.id",
-      "deleteRule": "id = @request.auth.id",
-      "options": {
-        "allowEmailAuth": true,
-        "allowOAuth2Auth": true,
-        "allowUsernameAuth": true,
-        "exceptEmailDomains": null,
-        "manageRule": null,
-        "minPasswordLength": 8,
-        "onlyEmailDomains": null,
-        "requireEmail": false
-      }
-    }
-  ];
-
-  const collections = snapshot.map((item) => new Collection(item));
-
-  return Dao(db).importCollections(collections, true, null);
-}, (db) => {
-  return null;
-})
