--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1685008757_updated_featureFlags.js
@@ -1,15 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("z4cc0g76ciqx13v")
-
-  collection.name = "features"
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("z4cc0g76ciqx13v")
-
-  collection.name = "featureFlags"
-
-  return dao.saveCollection(collection)
-})
