--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1686135066_updated_CrudExample_Events.js
@@ -1,23 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("2hkhhnkiep3qbyw")
-
-  collection.listRule = ""
-  collection.viewRule = ""
-  collection.createRule = ""
-  collection.updateRule = ""
-  collection.deleteRule = ""
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("2hkhhnkiep3qbyw")
-
-  collection.listRule = null
-  collection.viewRule = null
-  collection.createRule = null
-  collection.updateRule = null
-  collection.deleteRule = null
-
-  return dao.saveCollection(collection)
-})
