--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1686297520_updated_CrudExample_AllFields.js
@@ -1,51 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "3hmes6kr",
-    "name": "file",
-    "type": "file",
-    "required": false,
-    "unique": false,
-    "options": {
-      "maxSelect": 99,
-      "maxSize": 5242880,
-      "mimeTypes": [
-        "application/json",
-        "application/pdf"
-      ],
-      "thumbs": [],
-      "protected": false
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "3hmes6kr",
-    "name": "file",
-    "type": "file",
-    "required": false,
-    "unique": false,
-    "options": {
-      "maxSelect": 1,
-      "maxSize": 5242880,
-      "mimeTypes": [
-        "application/json",
-        "application/pdf"
-      ],
-      "thumbs": [],
-      "protected": false
-    }
-  }))
-
-  return dao.saveCollection(collection)
-})
