--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1686328805_updated_CrudExample_AllFields.js
@@ -1,41 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "o0hwn3yh",
-    "name": "text",
-    "type": "text",
-    "required": true,
-    "unique": false,
-    "options": {
-      "min": 6,
-      "max": 10,
-      "pattern": ""
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "o0hwn3yh",
-    "name": "text",
-    "type": "text",
-    "required": true,
-    "unique": false,
-    "options": {
-      "min": 6,
-      "max": 10,
-      "pattern": "^w+S"
-    }
-  }))
-
-  return dao.saveCollection(collection)
-})
