--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1686594012_updated_CrudExample_AllFields.js
@@ -1,15 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  collection.name = "CrudExample"
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  collection.name = "CrudExample_AllFields"
-
-  return dao.saveCollection(collection)
-})
