--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1687182078_updated_CrudExample.js
@@ -1,31 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // add
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "uy849pd2",
-    "name": "relation_single",
-    "type": "relation",
-    "required": true,
-    "unique": false,
-    "options": {
-      "collectionId": "z4cc0g76ciqx13v",
-      "cascadeDelete": false,
-      "minSelect": null,
-      "maxSelect": 1,
-      "displayFields": []
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // remove
-  collection.schema.removeField("uy849pd2")
-
-  return dao.saveCollection(collection)
-})
