--- saas/{{cookiecutter.project_name}}/admin/pb_migrations/1687182230_updated_CrudExample.js
@@ -1,51 +1,0 @@
-migrate((db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "nfghsb83",
-    "name": "image",
-    "type": "file",
-    "required": false,
-    "unique": false,
-    "options": {
-      "maxSelect": 1,
-      "maxSize": 5242880,
-      "mimeTypes": [
-        "image/png",
-        "image/jpeg"
-      ],
-      "thumbs": [],
-      "protected": false
-    }
-  }))
-
-  return dao.saveCollection(collection)
-}, (db) => {
-  const dao = new Dao(db)
-  const collection = dao.findCollectionByNameOrId("p9dhra382rts3bg")
-
-  // update
-  collection.schema.addField(new SchemaField({
-    "system": false,
-    "id": "nfghsb83",
-    "name": "image",
-    "type": "file",
-    "required": false,
-    "unique": false,
-    "options": {
-      "maxSelect": 99,
-      "maxSize": 5242880,
-      "mimeTypes": [
-        "image/png",
-        "image/jpeg"
-      ],
-      "thumbs": [],
-      "protected": false
-    }
-  }))
-
-  return dao.saveCollection(collection)
-})
